requirements.txt
smartapi-python>=1.3.4
pyotp>=2.8.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
websocket-client>=1.6.0
logzero>=1.7.0
